Implement Base Game Manager And Customer Classes
This commit adds:
WeaponEnum.cs: Holds enums representing aspects of the games weapons/weapon types.
Customer.cs: A class used to represent a customer. Holds functions to respond to round start and end (currently moving to counter location and end location). As well as holding a weapon request, a function to randomly generate a weapon request, and a function to display the request in the scene (not fully implemented, currently debug logs request). Finally a MoveTo(Vector3) function.
GameManager.cs: A class to handle general game flow. Holds delegates/events for phases (round start, round end) and manages the customers. Currently it spawns a customers and starts a round after a delay.

These classes contain ToDo: comments for the next steps.

Implementations of these can be found in Developers/Taisen/...