Implement Customer Movement Animations
This commit:
Creates an AnimationController: "AnimController_Customer" which moves the customer between an idle animation and a walk animation based on its current movement. Then a new prefab "Prefab_Customer_WithAnims" in "Taisen\TestPrefabs" was created with is a customer prefab with the animations setup.

GameManager.cs:
• Rotates spawned customer to face the counter at start.
• Adds small delay to spawning customer.

Class Implementations & Changes:
Customer.cs:
• Adds animator reference.
• Implements Update() to inform animator of movement.
• Adds reference to the player's camera.
• Implements a callback in MoveCustomerTo() to run a passed function when the customer arrives to the move location.
• Implemented OnReachedCounter() which rotates the customer towards the player camera and invokes DisplayCurrentWeaponRequestInWorld() after a short delay (delayBeforeRequestDisplay property).

This will be refactored in the next commit but saving implementation here.