Implement Customer Request World Space Speech Bubble
This commit implements creating/settings/displaying a weapon request speech bubble from our Customer script. I created a WorldSpaceTextBubble script to take a weapon type and change an image to the respective weapon type. Inside Customer.cs a weapon request speech bubble is created for the customer, then set for that instance's weapon request. The request bubble is displayed at the counter before being hidden after a delay.

Inside the editor, a prefab "Prefab_RequestSpeechBubble" was created. This is a world space canvas with the WorldSpaceTextBubble script on it. This prefab has a chat bubble with an image used to display a weapon type image. This prefab was set on our Prefab_Customer_WithAnims.

Note: I do not know what the Settings.json change was from, but I left it.

Prefab_Customer_WithAnims:
• Added empty GameObject for the location of weapon request bubble.

Prefab_RequestSpeechBubble:
• Created GameObject Canvas with speech bubble raw image and weapon raw image.
• Added WorldSpaceTextBubble script to it.

WorldSpaceTextBubble.cs:
• Holds reference to a raw image slot as well as the images for all the weapon types.
• public SetWeaponRequestImage(...), a function to set the weaponImageSlot to display an image based on the given WeaponType.

Customer.cs:
• Added properties for the weapon request bubble prefab, parent/spawn location, weapon request bubble instance, and a delay before the weapon request is hidden after displaying.
• Implemented creating and hiding the weapon request bubble in Start().
• Implemented DisplayCurrentWeaponRequestInWorld() to set our weapon request bubble's weapon image to the weapon request type. Then rotating the bubble to face the player camera, before revealing it. Then to finish off, hiding the request bubble after a delay.