Implement Smooth Rotation In Customer
This commit adds a coroutine SmoothlyRotateToTarget(Quaternion targetRotation) which smoothly interpolates rotation to the given target rotation. This coroutine is started in RotateCustomerToFaceTarget(Quaternion targetRotation). The call to set the customer's rotation directly in OnReachedCounter() was replaced with this implementation.

A property smoothRotationSpeed was also added to control said rotation speed.