(Saving) Add GameScore Struct
This commit adds a struct GameScore, which is used in handling the player's score. This struct holds properties used in calculating the score, and events/actions that are invoked when values are set.
An enum ScoreRankings was also added but it is not used anywhere yet.

NOTE:
This commit is just to save my progress. There are changes I am still going to make with the events. I just wanted to save this before I did them.

Current issues:
• I don't like my even handling. I created events for when a GameScore property is set, these are invoked through the setter, even if no changes are made and they pass the new value. Then I made a general GameScoreChanged event that passes the complete GameScore whenever a value is set and is changed (checks before and after). I am going to just have the single property events in the final version.
•