Implement A GameManager Callback Function For GameScore Updates
This commit adds a callback function OnGameScoreUpdated(..., ...) to the GameManager. This function receives the GameScoreProperty and Value that has been updated (set) and can respond. These responses are not implemented but ToDo comments were added.

The GameScore struct was also reorganized to group the members better. and an enum GameScoreProperty was added to represent the members of a GameScore. This enum was added to the GameScore events to and implemented in the callback function to more easily see what changes have occured.

Some renaming and all that was also done.